ai_web_tester/requirements.txt
# Core Framework
flask>=3.0.0
python-dotenv>=1.0.0

# LangGraph & LangChain
langgraph>=0.0.30
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0

# Playwright for browser automation
playwright>=1.40.0

# Streamlit for UI
streamlit>=1.30.0

# Utilities
pydantic>=2.0.0
